requirements.txt
# Data processing
pandas==2.2.3
numpy==2.2.6

# Web scraping & automation
selenium==4.31
PyYAML==6.0.2
PyMuPDF==1.26.0
python-docx==1.1.2

# Database handling
mysql-connector-python==9.3.0
SQLAlchemy==2.0.41

# Visualization
streamlit==1.45.1

# AI model inference
transformers==4.52.4
accelerate==0.27.2
git+https://github.com/bitsandbytes-foundation/bitsandbytes.git
#flash-attn==2.7.3
mistralai==1.8.1

# Also run the following
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121